# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tasklist API",
    "version": "1.4.2",
    "description": "Lists, tasks, due dates and completion. Every request needs a bearer token."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Local development" }],
  "security": [{ "bearerAuth": [] }],
  "tags": [
    { "name": "Lists", "description": "Named containers for tasks." },
    { "name": "Tasks", "description": "Things to do, each belonging to one list." }
  ],
  "paths": {
    "/lists": {
      "get": {
        "operationId": "listLists",
        "tags": ["Lists"],
        "summary": "List all lists",
        "responses": {
          "200": { "description": "All lists.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListCollection" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      },
      "post": {
        "operationId": "createList",
        "tags": ["Lists"],
        "summary": "Create a list",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewList" } } } },
        "responses": {
          "201": { "description": "The created list.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/List" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/lists/{listId}": {
      "parameters": [{ "$ref": "#/components/parameters/ListId" }],
      "get": {
        "operationId": "getList",
        "tags": ["Lists"],
        "summary": "Get a list",
        "responses": {
          "200": { "description": "The list.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/List" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteList",
        "tags": ["Lists"],
        "summary": "Delete a list and its tasks",
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/tasks": {
      "get": {
        "operationId": "listTasks",
        "tags": ["Tasks"],
        "summary": "List tasks",
        "parameters": [
          { "name": "listId", "in": "query", "description": "Only tasks in this list.", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "description": "Only tasks with this status.", "schema": { "$ref": "#/components/schemas/TaskStatus" } }
        ],
        "responses": {
          "200": { "description": "Matching tasks.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskCollection" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      },
      "post": {
        "operationId": "createTask",
        "tags": ["Tasks"],
        "summary": "Create a task",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewTask" } } } },
        "responses": {
          "201": { "description": "The created task.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/tasks/{taskId}": {
      "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
      "get": {
        "operationId": "getTask",
        "tags": ["Tasks"],
        "summary": "Get a task",
        "responses": {
          "200": { "description": "The task.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateTask",
        "tags": ["Tasks"],
        "summary": "Update a task",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskUpdate" } } } },
        "responses": {
          "200": { "description": "The updated task.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "tags": ["Tasks"],
        "summary": "Delete a task",
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/tasks/{taskId}/complete": {
      "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
      "post": {
        "operationId": "completeTask",
        "tags": ["Tasks"],
        "summary": "Mark a task as done",
        "responses": {
          "200": { "description": "The completed task.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "The task was already done.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Any API token issued from the account settings page." }
    },
    "parameters": {
      "ListId": { "name": "listId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "lst_inbox" },
      "TaskId": { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "tsk_01" }
    },
    "responses": {
      "Unauthenticated": { "description": "The bearer token is missing or malformed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "No resource with that id.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "ValidationError": { "description": "The request body or query is invalid.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "TaskStatus": { "type": "string", "enum": ["open", "done"] },
      "List": {
        "type": "object",
        "required": ["id", "name", "createdAt"],
        "properties": {
          "id": { "type": "string", "example": "lst_inbox" },
          "name": { "type": "string", "example": "Inbox" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "NewList": {
        "type": "object",
        "required": ["name"],
        "properties": { "name": { "type": "string", "minLength": 1, "example": "Launch checklist" } }
      },
      "ListCollection": {
        "type": "object",
        "required": ["data"],
        "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/List" } } }
      },
      "Task": {
        "type": "object",
        "required": ["id", "listId", "title", "notes", "dueDate", "status", "createdAt", "completedAt"],
        "properties": {
          "id": { "type": "string", "example": "tsk_01" },
          "listId": { "type": "string", "example": "lst_inbox" },
          "title": { "type": "string", "example": "Renew domain" },
          "notes": { "type": "string", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true, "example": "2026-02-01" },
          "status": { "$ref": "#/components/schemas/TaskStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["listId", "title"],
        "properties": {
          "listId": { "type": "string", "example": "lst_inbox" },
          "title": { "type": "string", "minLength": 1, "example": "Renew domain" },
          "notes": { "type": "string" },
          "dueDate": { "type": "string", "format": "date" }
        }
      },
      "TaskUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "notes": { "type": "string", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "TaskCollection": {
        "type": "object",
        "required": ["data"],
        "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "not_found" },
              "message": { "type": "string", "example": "No task with that id." }
            }
          }
        }
      }
    }
  }
}